Fixed Easter egg's memory loss

+++ Forms/SettingsForm.cs
@@ -247,8 +247,9 @@ namespace SteamVR_OculusDash_Switcher
 					{
 						picRealismLevel.BackgroundImage = GetImage("Settings", "RealismLevelImage", "Max"); //: add meme image left to that slider
 
-						File.Copy(@"images\Settings\Max\Horizon\Heart.png", @"images\Settings\Max\Horizon\Heart.mp3");
-						Egg = new WindowsMediaPlayer(); ;
+						if (File.Exists(@"images\Settings\Max\Horizon\Heart.mp3"))
+							File.Copy(@"images\Settings\Max\Horizon\Heart.png", @"images\Settings\Max\Horizon\Heart.mp3");
+						Egg = new WindowsMediaPlayer();
 						Egg.URL = @"images\Settings\Max\Horizon\Heart.mp3";
 						Egg.controls.play();//: play BMttH: can you break my heart
 					}
@@ -257,13 +258,18 @@ namespace SteamVR_OculusDash_Switcher
 						Egg?.controls.stop();
 						if(File.Exists(@"images\Settings\Max\Horizon\Heart.mp3")) 
 							File.Delete(@"images\Settings\Max\Horizon\Heart.mp3");
+						Egg.close();
 						Egg = null;
 						exception.ShowMessageBox(LocalizationStrings.SettingsForm_RealismLevel_Max_Error);
 					}
 				}
 				else
 				{
+					if(File.Exists(@"images\Settings\Max\Horizon\Heart.mp3")) 
+						File.Delete(@"images\Settings\Max\Horizon\Heart.mp3");
+
 					Egg?.controls.stop();
+					Egg.close();
 					Egg = null;
 				}
 				Settings.Default.Save();
